requirements.txt
julep==1.48.1
python-dotenv==1.0.1
redis==5.0.1
chromadb==0.4.22
sentence-transformers==2.5.1
python-magic==0.4.27
PyYAML==6.0.1
pydantic>=1.9.0
httpx>=0.23.0
ruamel-yaml>=0.18.6
